tests/user_service/integration_tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiokafka==0.12.0
PyJWT==2.8.0
python-multipart==0.0.6
coverage==7.3.2
pytest-cov==4.1.0
requests==2.31.0